+++ CHANGELOG.md
@@ -4,10 +4,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 The format conforms to [Keep a Changelog](http://keepachangelog.com/).
 
 ## [Unreleased]
-### Changed
-- Allow `allowShortCircuit` and `allowTernary` for `no-unused-expressions`
 ### Added
 - Added `max-len` rule
+- Added `max: 1`, `maxEOF: 0` and `maxBOF: 0` to `no-multiple-empty-lines`
+### Changed
+- Allow `allowShortCircuit` and `allowTernary` for `no-unused-expressions`
 
 ## [1.1.0] - 2016-06-17
 ### Changed